requirements.txt
aerich
aiofiles
aiosqlite
aniso8601
asgiref
async-exit-stack
async-generator
asyncpg
bcrypt
certifi
cffi
charset-normalizer
click
cryptography
ddlparse
dictdiffer
dnspython
ecdsa
email-validator
fastapi
graphene
graphql-core
graphql-relay
gunicorn
h11
httptools
idna
iso8601
itsdangerous
Jinja2
MarkupSafe
orjson
passlib
prodict
promise
pyasn1
pycparser
pydantic
PyJWT
pyparsing
pypika-tortoise
python-dotenv
python-http-client
python-multipart
pytz
PyYAML
requests
rsa
Rx
sendgrid
six
starkbank-ecdsa
starlette
tortoise-orm
typing-extensions
ujson
urllib3
uvicorn
uvloop
watchgod
websockets
socketio
python-socketio